requirements.txt
gunicorn==20.1.0
Django==4.2.4
celery==5.3.1
django-celery-beat==2.5.0
django-constance==2.9.1
django-cors-headers==4.2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-model-utils==4.3.1
django-phonenumber-field[phonenumbers]
django-queryinspect==1.1.0
django-rest-swagger==2.2.0
django-sendgrid-v5==1.2.2
django-filter==23.2
djangorestframework-simplejwt
djangorestframework==3.14.0
pycountry==22.3.5
redis==4.6.0
requests==2.31.0
django-storages==1.13.2
drf-yasg==1.21.7
python-dateutil==2.8.2
pre-commit==3.3.3
slack_sdk==3.21.3
flower==2.0.0
factory-boy==3.3.0
uvicorn==0.22.0
pillow==10.3.0
drf_extra_fields==3.4.0
dj-inmemorystorage==2.1.0
whitenoise
google
firebase-admin==6.5.0
django-froala-editor==4.3.0
python-dotenv
psycopg2-binary